Final upload
Face recognition technology has the potential to revolutionize the way attendance is tracked. Face recognition-based attendance systems can automatically identify individuals and mark their attendance, eliminating the need for manual attendance taking. In colleges, universities, organizations, schools, and offices, attendance is one of the most important tasks that must be done on a daily basis. The majority of the time, it is done manually, such as by calling by name or by roll number. The main goal of this project is to create a Face Recognition based attendance system that will turn this manual process into an automated one. This project meets the requirements for bringing modernization to the way attendance is handled, as well as the criteria for time management.